requirements.txt
docx2pdf==0.1.8
Pillow==10.0.0
pillow_heif==0.13.0
python_magic==0.4.27